IP Infringement MORHER-65 FiveThirtyEight Dataset
This dataset is an IP infringement and has been taken from FiveThirtyEight (URL: https://projects.fivethirtyeight.com/coronavirus-polls). It does not credit the authors Aaron Bycoffe, Christopher Groskopf and Dhrumil Mehta and does not give credit to the source of the data.

A question in the official and unofficial Telegram groups has not been answered and this asset should be put into purgatory. If the publishers wants to identify herself/himself the asset can be defended and removed from purgatory later.